+++ flutter/test/integrations/screenshot_integration_test.dart
@@ -1,7 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 import 'package:sentry_flutter/src/integrations/screenshot_integration.dart';
-import 'package:sentry_flutter/src/integrations/widgets_binding_integration.dart';
 import 'package:sentry_flutter/src/screenshot/screenshot_attachment_processor.dart';
 
 import '../mocks.mocks.dart';