+++ flutter/ios/Classes/SentryFlutterPluginApple.swift
@@ -527,7 +527,7 @@ public class SentryFlutterPluginApple: NSObject, FlutterPlugin {
           }
         }
 
-        // TODO: missing name and geo
+        // missing name and geo
         // Should be solved by https://github.com/getsentry/team-mobile/issues/59
         // or https://github.com/getsentry/team-mobile/issues/56